+++ src/OpenTelemetry.AutoInstrumentation/Configurations/Otlp/OtlpSignalType.cs
@@ -1,0 +1,11 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.AutoInstrumentation.Configurations.Otlp;
+
+internal enum OtlpSignalType
+{
+    Traces,
+    Metrics,
+    Logs
+}